Update: move ft_print_program_struct func
